printing_service/requirements.txt
bleak>=0.20.0
Pillow>=10.0.0
asyncio>=3.4.3
python-dotenv>=1.0.0